+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/message/internal/MessageDao.kt
@@ -29,10 +29,10 @@ import java.util.Date
 
 @Dao
 @Suppress("TooManyFunctions")
-internal abstract class MessageDao {
+internal interface MessageDao {
 
     @Transaction
-    open suspend fun insert(messageEntities: List<MessageEntity>) {
+    suspend fun insert(messageEntities: List<MessageEntity>) {
         upsertMessageInnerEntities(messageEntities.map(MessageEntity::messageInnerEntity))
         deleteAttachments(messageEntities.map { it.messageInnerEntity.id })
         insertAttachments(messageEntities.flatMap(MessageEntity::attachments))
@@ -40,27 +40,27 @@ internal abstract class MessageDao {
     }
 
     @Transaction
-    open fun deleteAttachments(messageIds: List<String>) {
+    fun deleteAttachments(messageIds: List<String>) {
         messageIds.chunked(SQLITE_MAX_VARIABLE_NUMBER).forEach(::deleteAttachmentsChunked)
     }
 
     @Query("DELETE FROM attachment_inner_entity WHERE messageId in (:messageIds)")
-    protected abstract fun deleteAttachmentsChunked(messageIds: List<String>)
+    fun deleteAttachmentsChunked(messageIds: List<String>)
 
     @Transaction
-    open suspend fun insert(messageEntity: MessageEntity) = insert(listOf(messageEntity))
+    suspend fun insert(messageEntity: MessageEntity) = insert(listOf(messageEntity))
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
-    protected abstract suspend fun insertMessageInnerEntity(messageInnerEntity: MessageInnerEntity): Long
+    suspend fun insertMessageInnerEntity(messageInnerEntity: MessageInnerEntity): Long
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
-    protected abstract suspend fun insertMessageInnerEntities(messageInnerEntities: List<MessageInnerEntity>): List<Long>
+    suspend fun insertMessageInnerEntities(messageInnerEntities: List<MessageInnerEntity>): List<Long>
 
     @Update(onConflict = OnConflictStrategy.REPLACE)
-    protected abstract fun updateMessageInnerEntity(messageInnerEntity: MessageInnerEntity)
+    fun updateMessageInnerEntity(messageInnerEntity: MessageInnerEntity)
 
     @Transaction
-    open suspend fun upsertMessageInnerEntity(messageInnerEntity: MessageInnerEntity) {
+    suspend fun upsertMessageInnerEntity(messageInnerEntity: MessageInnerEntity) {
         val rowId = insertMessageInnerEntity(messageInnerEntity)
         if (rowId == -1L) {
             updateMessageInnerEntity(messageInnerEntity)
@@ -68,7 +68,7 @@ internal abstract class MessageDao {
     }
 
     @Transaction
-    open suspend fun upsertMessageInnerEntities(messageInnerEntities: List<MessageInnerEntity>) {
+    suspend fun upsertMessageInnerEntities(messageInnerEntities: List<MessageInnerEntity>) {
         val rowIds = insertMessageInnerEntities(messageInnerEntities)
         val entitiesToUpdate = rowIds.mapIndexedNotNull { index, rowId ->
             if (rowId == -1L) messageInnerEntities[index] else null
@@ -77,87 +77,137 @@ internal abstract class MessageDao {
     }
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    protected abstract suspend fun insertAttachments(attachmentEntities: List<AttachmentEntity>)
+    suspend fun insertAttachments(attachmentEntities: List<AttachmentEntity>)
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    protected abstract suspend fun insertReactions(reactions: List<ReactionEntity>)
+    suspend fun insertReactions(reactions: List<ReactionEntity>)
 
-    @Query("SELECT * from stream_chat_message WHERE cid = :cid AND (createdAt > :dateFilter || createdLocallyAt > :dateFilter) ORDER BY CASE WHEN createdAt IS NULL THEN createdLocallyAt ELSE createdAt END ASC LIMIT :limit")
+    @Query(
+        "SELECT * from $MESSAGE_ENTITY_TABLE_NAME " +
+            "WHERE cid = :cid " +
+            "AND (createdAt > :dateFilter || createdLocallyAt > :dateFilter) " +
+            "ORDER BY CASE WHEN createdAt " +
+            "IS NULL THEN createdLocallyAt " +
+            "ELSE createdAt " +
+            "END ASC LIMIT :limit"
+    )
     @Transaction
-    abstract suspend fun messagesForChannelNewerThan(
+    suspend fun messagesForChannelNewerThan(
         cid: String,
         limit: Int = 100,
         dateFilter: Date,
     ): List<MessageEntity>
 
-    @Query("SELECT * from stream_chat_message WHERE cid = :cid AND (createdAt >= :dateFilter || createdLocallyAt >= :dateFilter) ORDER BY CASE WHEN createdAt IS NULL THEN createdLocallyAt ELSE createdAt END ASC LIMIT :limit")
+    @Query(
+        "SELECT * from $MESSAGE_ENTITY_TABLE_NAME " +
+            "WHERE cid = :cid " +
+            "AND (createdAt >= :dateFilter || createdLocallyAt >= :dateFilter) " +
+            "ORDER BY CASE WHEN createdAt " +
+            "IS NULL THEN createdLocallyAt " +
+            "ELSE createdAt " +
+            "END ASC LIMIT :limit"
+    )
     @Transaction
-    abstract suspend fun messagesForChannelEqualOrNewerThan(
+    suspend fun messagesForChannelEqualOrNewerThan(
         cid: String,
         limit: Int = 100,
         dateFilter: Date,
     ): List<MessageEntity>
 
-    @Query("SELECT * from stream_chat_message WHERE cid = :cid AND (createdAt < :dateFilter || createdLocallyAt < :dateFilter) ORDER BY CASE WHEN createdAt IS NULL THEN createdLocallyAt ELSE createdAt END DESC LIMIT :limit")
+    @Query(
+        "SELECT * from $MESSAGE_ENTITY_TABLE_NAME " +
+            "WHERE cid = :cid " +
+            "AND (createdAt < :dateFilter || createdLocallyAt < :dateFilter) " +
+            "ORDER BY CASE WHEN createdAt " +
+            "IS NULL THEN createdLocallyAt " +
+            "ELSE createdAt " +
+            "END DESC LIMIT :limit"
+    )
     @Transaction
-    abstract suspend fun messagesForChannelOlderThan(
+    suspend fun messagesForChannelOlderThan(
         cid: String,
         limit: Int = 100,
         dateFilter: Date,
     ): List<MessageEntity>
 
-    @Query("SELECT * from stream_chat_message WHERE cid = :cid AND (createdAt <= :dateFilter || createdLocallyAt <= :dateFilter) ORDER BY CASE WHEN createdAt IS NULL THEN createdLocallyAt ELSE createdAt END DESC LIMIT :limit")
+    @Query(
+        "SELECT * from $MESSAGE_ENTITY_TABLE_NAME " +
+            "WHERE cid = :cid " +
+            "AND (createdAt <= :dateFilter || createdLocallyAt <= :dateFilter) " +
+            "ORDER BY CASE WHEN createdAt " +
+            "IS NULL THEN createdLocallyAt " +
+            "ELSE createdAt " +
+            "END DESC LIMIT :limit"
+    )
     @Transaction
-    abstract suspend fun messagesForChannelEqualOrOlderThan(
+    suspend fun messagesForChannelEqualOrOlderThan(
         cid: String,
         limit: Int = 100,
         dateFilter: Date,
     ): List<MessageEntity>
 
-    @Query("SELECT * from stream_chat_message WHERE cid = :cid ORDER BY CASE WHEN createdAt IS NULL THEN createdLocallyAt ELSE createdAt END DESC LIMIT :limit")
+    @Query(
+        "SELECT * from $MESSAGE_ENTITY_TABLE_NAME " +
+            "WHERE cid = :cid " +
+            "ORDER BY CASE WHEN createdAt " +
+            "IS NULL THEN createdLocallyAt " +
+            "ELSE createdAt " +
+            "END DESC LIMIT :limit"
+    )
     @Transaction
-    abstract suspend fun messagesForChannel(cid: String, limit: Int = 100): List<MessageEntity>
+    suspend fun messagesForChannel(cid: String, limit: Int = 100): List<MessageEntity>
 
-    @Query("DELETE from stream_chat_message WHERE cid = :cid AND createdAt < :deleteMessagesBefore")
-    abstract suspend fun deleteChannelMessagesBefore(cid: String, deleteMessagesBefore: Date)
+    @Query(
+        "DELETE from $MESSAGE_ENTITY_TABLE_NAME " +
+            "WHERE cid = :cid " +
+            "AND createdAt < :deleteMessagesBefore"
+    )
+    suspend fun deleteChannelMessagesBefore(cid: String, deleteMessagesBefore: Date)
 
-    @Query("DELETE from stream_chat_message WHERE cid = :cid AND id = :messageId")
-    abstract suspend fun deleteMessage(cid: String, messageId: String)
+    @Query(
+        "DELETE from $MESSAGE_ENTITY_TABLE_NAME " +
+            "WHERE cid = :cid " +
+            "AND id = :messageId"
+    )
+    suspend fun deleteMessage(cid: String, messageId: String)
 
     @Transaction
-    open suspend fun select(ids: List<String>): List<MessageEntity> {
+    suspend fun select(ids: List<String>): List<MessageEntity> {
         return ids.chunked(SQLITE_MAX_VARIABLE_NUMBER).flatMap { messageIds -> selectChunked(messageIds) }
     }
 
-    @Query("SELECT * FROM stream_chat_message WHERE id IN (:ids)")
+    @Query("SELECT * FROM $MESSAGE_ENTITY_TABLE_NAME WHERE id IN (:ids)")
     @Transaction
-    protected abstract suspend fun selectChunked(ids: List<String>): List<MessageEntity>
+    suspend fun selectChunked(ids: List<String>): List<MessageEntity>
 
-    @Query("SELECT * FROM stream_chat_message WHERE id IN (:id)")
+    @Query("SELECT * FROM $MESSAGE_ENTITY_TABLE_NAME WHERE id IN (:id)")
     @Transaction
-    abstract suspend fun select(id: String): MessageEntity?
+    suspend fun select(id: String): MessageEntity?
 
     @Transaction
-    open suspend fun selectWaitForAttachments(): List<MessageEntity> {
+    suspend fun selectWaitForAttachments(): List<MessageEntity> {
         return selectBySyncStatus(SyncStatus.AWAITING_ATTACHMENTS)
     }
 
     @Query(
-        "SELECT * FROM stream_chat_message " +
+        "SELECT * FROM $MESSAGE_ENTITY_TABLE_NAME " +
             "WHERE syncStatus = :syncStatus " +
             "ORDER BY CASE WHEN createdAt IS NULL THEN createdLocallyAt ELSE createdAt END ASC " +
             "LIMIT :limit"
     )
     @Transaction
-    abstract suspend fun selectBySyncStatus(syncStatus: SyncStatus, limit: Int = NO_LIMIT): List<MessageEntity>
+    suspend fun selectBySyncStatus(syncStatus: SyncStatus, limit: Int = NO_LIMIT): List<MessageEntity>
 
     @Query(
-        "SELECT id FROM stream_chat_message " +
+        "SELECT id FROM $MESSAGE_ENTITY_TABLE_NAME " +
             "WHERE syncStatus = :syncStatus " +
             "ORDER BY CASE WHEN createdAt IS NULL THEN createdLocallyAt ELSE createdAt END ASC " +
             "LIMIT :limit"
     )
-    abstract suspend fun selectIdsBySyncStatus(syncStatus: SyncStatus, limit: Int = NO_LIMIT): List<String>
+    suspend fun selectIdsBySyncStatus(syncStatus: SyncStatus, limit: Int = NO_LIMIT): List<String>
+
+    @Query("DELETE FROM $MESSAGE_ENTITY_TABLE_NAME")
+    suspend fun deleteAll()
 
     private companion object {
         private const val SQLITE_MAX_VARIABLE_NUMBER: Int = 999