+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/reaction/internal/ReactionDao.kt
@@ -28,14 +28,14 @@ internal interface ReactionDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(reactionEntity: ReactionEntity)
 
-    @Query("SELECT * FROM stream_chat_reaction WHERE id = :id")
+    @Query("SELECT * FROM $REACTION_ENTITY_TABLE_NAME WHERE id = :id")
     suspend fun selectReactionById(id: Int): ReactionEntity?
 
-    @Query("SELECT * FROM stream_chat_reaction WHERE id IN (:ids)")
+    @Query("SELECT * FROM $REACTION_ENTITY_TABLE_NAME WHERE id IN (:ids)")
     suspend fun selectReactionsByIds(ids: List<Int>): List<ReactionEntity>
 
     @Query(
-        "SELECT id FROM stream_chat_reaction " +
+        "SELECT id FROM $REACTION_ENTITY_TABLE_NAME " +
             "WHERE syncStatus = :syncStatus " +
             "ORDER BY syncStatus ASC " +
             "LIMIT :limit"
@@ -50,14 +50,31 @@ internal interface ReactionDao {
     )
     suspend fun selectSyncStatus(syncStatus: SyncStatus, limit: Int = NO_LIMIT): List<ReactionEntity>
 
-    @Query("SELECT * FROM stream_chat_reaction WHERE stream_chat_reaction.type = :reactionType AND stream_chat_reaction.messageid = :messageId AND userId = :userId")
+    @Query(
+        "SELECT * FROM $REACTION_ENTITY_TABLE_NAME " +
+            "WHERE stream_chat_reaction.type = :reactionType " +
+            "AND stream_chat_reaction.messageid = :messageId " +
+            "AND userId = :userId"
+    )
     suspend fun selectUserReactionToMessage(reactionType: String, messageId: String, userId: String): ReactionEntity?
 
-    @Query("SELECT * FROM stream_chat_reaction WHERE stream_chat_reaction.messageid = :messageId AND userId = :userId")
+    @Query(
+        "SELECT * FROM $REACTION_ENTITY_TABLE_NAME " +
+            "WHERE stream_chat_reaction.messageid = :messageId " +
+            "AND userId = :userId"
+    )
     suspend fun selectUserReactionsToMessage(messageId: String, userId: String): List<ReactionEntity>
 
-    @Query("UPDATE stream_chat_reaction SET deletedAt = :deletedAt WHERE userId = :userId AND messageId = :messageId")
+    @Query(
+        "UPDATE $REACTION_ENTITY_TABLE_NAME " +
+            "SET deletedAt = :deletedAt " +
+            "WHERE userId = :userId " +
+            "AND messageId = :messageId"
+    )
     suspend fun setDeleteAt(userId: String, messageId: String, deletedAt: Date)
+
+    @Query("DELETE FROM $REACTION_ENTITY_TABLE_NAME")
+    suspend fun deleteAll()
 
     private companion object {
         private const val NO_LIMIT: Int = -1