+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/user/internal/UserDao.kt
@@ -30,15 +30,25 @@ internal interface UserDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(userEntity: UserEntity)
 
-    @Query("SELECT * FROM stream_chat_user WHERE stream_chat_user.id IN (:ids)")
+    @Query("SELECT * FROM $USER_ENTITY_TABLE_NAME WHERE stream_chat_user.id IN (:ids)")
     suspend fun select(ids: List<String>): List<UserEntity>
 
-    @Query("SELECT * FROM stream_chat_user WHERE stream_chat_user.id IN (:id)")
+    @Query("SELECT * FROM $USER_ENTITY_TABLE_NAME WHERE stream_chat_user.id IN (:id)")
     suspend fun select(id: String): UserEntity?
 
-    @Query("SELECT * FROM stream_chat_user ORDER BY name ASC LIMIT :limit OFFSET :offset")
+    @Query("SELECT * FROM $USER_ENTITY_TABLE_NAME ORDER BY name ASC LIMIT :limit OFFSET :offset")
     fun selectAllUser(limit: Int, offset: Int): List<UserEntity>
 
-    @Query("SELECT * FROM stream_chat_user WHERE name LIKE :searchString ORDER BY name ASC LIMIT :limit OFFSET :offset")
+    @Query(
+        "SELECT * FROM $USER_ENTITY_TABLE_NAME " +
+            "WHERE name " +
+            "LIKE :searchString " +
+            "ORDER BY name " +
+            "ASC LIMIT :limit " +
+            "OFFSET :offset"
+    )
     fun selectUsersLikeName(searchString: String, limit: Int, offset: Int): List<UserEntity>
+
+    @Query("DELETE FROM $USER_ENTITY_TABLE_NAME")
+    suspend fun deleteAll()
 }